+++ package.json
@@ -15,6 +15,7 @@
     "@fortawesome/fontawesome-svg-core": "^6.2.0",
     "@fortawesome/free-solid-svg-icons": "^6.2.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mui/icons-material": "^5.11.9",
     "@mui/material": "^5.10.12",
     "@types/node": "18.11.9",
     "@types/react": "18.0.24",
