+++ src/pages/index.tsx
@@ -8,19 +8,9 @@ import { auth } from '../services/firebase'
 import styles from '../styles/Home.module.css'
 
 export default function Home() {
-  const [data, setData] = useState<object>({})
-  const { session, setSession } = useContext(AppContext)
-  useEffect(() => {
-    setData(session)
-  })
-  const handleLogin = () => {
-    const provider = new GoogleAuthProvider()
-    provider.setCustomParameters({hd: "gct.ac.in"})
-    signInWithPopup(auth, provider).then(res => setSession(res.user))
-  }
   return (
-   <div>
-    {Object.keys(data).length==0 ? (<Button className="text-error-main" onClick={handleLogin}> Hello </Button>) : (<p>Logged in</p>)}
-   </div>
+    <div>
+      <h1>Hello World! This is the Home page</h1>
+    </div>
   )
 }
