+++ src/theme/theme.tsx
@@ -1,0 +1,35 @@
+import { createTheme } from "@mui/system";
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#f1b900',
+    },
+    // secondary: {
+    //   main: '#6c07ad',
+    //   light: '#9847cb',
+    //   dark: '#501b73',
+    // },
+    // error: {
+    //   main: '#ff1100',
+    //   dark: '#9e1d15',
+    // },
+    // warning: {
+    //   main: '#ff9700',
+    //   light: '#fbad3d',
+    //   dark: '#af6a07',
+    // },
+    // info: {
+    //   main: '#2596f1',
+    //   light: '#55aff7',
+    //   dark: '#195d92',
+    // },
+    // success: {
+    //   main: '#47b34b',
+    //   light: '#73c377',
+    //   dark: '#275e29',
+    // },
+  }
+})
+
+export default theme
